Move non-Cartesian3D benchmarks out of Cartesian3D.
